TASK: Draft for UI
see https://github.com/mstruebing/PackageFactory.Guevara/pull/2/files

Current state: :(
React.createElement: type is invalid -- expected a string (for built-in components) or a class/function (for composite components) but got: object.
Error: Element type is invalid: expected a string (for built-in components) or a class/function (for composite components) but got: object.